main.go: Read config from a new format TOML file
Construct a new config format, written in TOML. Read that format in
when starting the program. This new format has the benefit of using
project name aliases as keys. The goal will be to allow users to send
one of those aliases as a command line argument to the program,
and thus to have the program post a TimeTask entry for that project.

Here's an idea of what the new format looks like:

    [auth]
    username = "example"
    password_cmd = ""

    [projects.myprojectalias]
    client = ...
    project = ...
    module = ...
    task = ...
    work_type = ...
    time = 7
    billable = true

    [projects.project2]
    client = ...
    project = ...
    module = ...
    task = ...
    work_type = ...
    time = 7
    billable = true

Eventually, we'll need to remove the `interface{}` from the
`Projects` map value and replace it with a real type, but this was just
to test that it was possible to get us a nice map from the TOML.