Comment out all the things!
Half get rid of a lot of code. I don't like and don't want to use our
old field types. Get rid of them and the code in 'http.go' that depends
on them.

Also get rid of the time entry submission code in 'main.go' as that's
going to be redone.