Remove generator.go, templates/ directory
These are not needed in the new world where we only submit a single time
entry at a time and entry parameters are filled in on the command line.

My guess is we'll likely have some sort of generator in the future to
create the initial `config.toml` file, but that won't contain any
logic-based data. My guess is we'll probably be able to just stick it in
a template string right inside a *.go file.