main.go: Create time entry before making HTTP requests
Make it easier to comment out the HTTP requests when testing command
line argument parsing. To test, we're just printing the time entry with:

    log.Printf("%+v\n", time_entry)

Fill in the values from our new command line argument variables. These
need to be dereferenced as they come out of Kingpin as pointers.

Haven't completely worked out how to deal with the 'date' argument yet
so leaving it for later.

Need to fail early if `project_alias` isn't recorgnised.