buildSubmissionParams(): Fix time string replacement bug
We ran into a bug where submitting "7" as the time spent would become
"700" on the site.

This was because our `n` value of 0 wasn't replacing the "." in the
string. Thus "7.00" became "7.00" after the replacement. Not right. I
misunderstood what that argument was doing and what the word "empty" in
the documentation meant.

Change the `n` value to replace all "."s in the string.