Add `MaybeWriteConfig()`
A new function that will write a new config.toml file to the
XDG_CONFIG_HOME directory. Currently it checks to see whether our config
file is present. If not and our config directory isn't present, it
creates it.

Still need to get this to actually write the config file.

Also, we won't want to call it by default in main() like we're doing
now. Will likely want to hide it behind a `--write-config` flag.