Add module.go for parsing Module XML
A new function that parses the XML returned by the `RequestModules()`
function.

It provides a `Module` type that allows us to interact with modules more
easily in code.

The `ParseXML()` function will take an XML string and return a slice of
`Module`s.

Added a test just to facilitate development. Wasn't able to find an easy
way to compare slices in Go, so just printed the values and checked the
result visually. Not a useful test for future use, but it served its
purpose. Eventually it would be nice to find a way to compare structs
and have a real pass/fail condition.