RequestModules(): Pretty print modules
Parse the module XML with `ModuleParseXML()`. Take the resulting
`[]Module` slice and use it to generate a string of the following
format:

    ID	Module
    55555	R&amp;D
    77777	Sprint 1
    222222 Sprint 2

This string is what gets printed to the console, which makes it rather
easy to read the modules that are available for the given project and
grab the appropriate ID to put into your config file.